requirements.txt
fastapi
pydantic
openai
python-dotenv
PyGithub
uvicorn
supabase